Make widget display correctly for singular posts
Singular posts weren't rendering correctly because inside the first if block condition, `$post` was used, but `$post` is not initialized at that point. Correct variable here would've been `$posts[0]`.

I figured that a singular post should not look any different to 5 posts in this widget, so removed the condition all together. But maybe I'm missing some edge case someone is getting using it like this?